requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==18.6.2
Pillow==9.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
